Added adding emojis
With the use of :name_of_emoji: user now has the ability to add emojis
that can be found in the resources folder. I've excluded the images in
this resource folder, but it is noted on the readme. I've also set the
gitignore to not care about them. I've also remaed the message commands
to make it seem more consistent with everything. Hopefully this isn't
confusing