Intial commit for XKCD command
This is an intial commit for the xkcd command that I'm adding. At the
moment this command only sends the most recent command it doesn't
actually even work correctly. I am only committing this because it's at
a point where it overall works, it's just going to need to have a lot of
work with documentation and fixing a few things like requesting any
xkcd.