Breaking up onMessageReceived
So, broke up onMessageReceived and such in order to try and organize it
a little bit better. This required adding event to ParsedCommandMessage
and we might need to have a ParsedText or something along those lines as
well. We also have added a lot more to try and build up logging messages
in output and such.