Refactoring + Changing how watch works
Watch can now watch certain people with persistence. The same for watch
all. That may change in the future, but at the moment that's how it is.

Most of the run commands have now been change to have
runWithArguments and runWithoutArguments respectively in order to deal a
little bit better with arguments

arguments have been changed to be a list of arguments and each one of
them is contained in their own String array...if that makes sense.